requirements.txt
# ============================================================================
# RiceGuard Backend Requirements (FastAPI + MongoDB + ML)
# Python 3.13 Compatible - Latest Stable Versions
# ============================================================================

# ----------------------------------------------------------------------------
# Core web stack - Latest stable versions with Python 3.13 support
# ----------------------------------------------------------------------------
fastapi==0.115.6
uvicorn[standard]==0.32.1
starlette>=0.41.3

# ----------------------------------------------------------------------------
# Validation & request parsing - Pydantic v2 ecosystem
# ----------------------------------------------------------------------------
pydantic[email]==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
python-multipart==0.0.17

# ----------------------------------------------------------------------------
# Configuration / environment handling
# ----------------------------------------------------------------------------
python-dotenv==1.0.1

# ----------------------------------------------------------------------------
# Database (MongoDB Atlas compatible) - Latest stable versions
# ----------------------------------------------------------------------------
pymongo[srv]==4.10.1
dnspython==2.7.0

# ----------------------------------------------------------------------------
# Security / authentication - Updated versions
# ----------------------------------------------------------------------------
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
cryptography==44.0.0

# ----------------------------------------------------------------------------
# Machine learning / image processing - Python 3.13 compatible
# ----------------------------------------------------------------------------
# TensorFlow doesn't support Python 3.13 yet, use environment markers
tensorflow==2.18.0 ; python_version < "3.13"
# For Python 3.13+, consider using torch/opencv or wait for TensorFlow support
numpy==1.26.4 ; python_version < "3.13"
numpy>=2.1.0,<3.0 ; python_version >= "3.13"
pillow==11.0.0
opencv-python==4.10.0.84
scikit-learn==1.5.2

# ----------------------------------------------------------------------------
# Observability / utilities - Updated versions
# ----------------------------------------------------------------------------
requests==2.32.3
typing-extensions>=4.12.2
certifi>=2024.8.30
urllib3==2.3.0
charset-normalizer==3.4.1

# ----------------------------------------------------------------------------
# File handling and additional utilities
# ----------------------------------------------------------------------------
aiofiles==24.1.0
python-magic==0.4.27

# ----------------------------------------------------------------------------
# Development and testing dependencies - Latest versions
# ----------------------------------------------------------------------------
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.1
pre-commit==4.0.1

# ----------------------------------------------------------------------------
# Performance and monitoring (optional but recommended)
# ----------------------------------------------------------------------------
redis==5.2.1
structlog==24.4.0

# ----------------------------------------------------------------------------
# API documentation and OpenAPI support
# ----------------------------------------------------------------------------
fastapi-users[oauth]==13.0.0  # For enhanced auth features
python-multipart==0.0.17  # Already listed above but required for file uploads
